% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compriskMPL-package.R
\docType{package}
\name{compriskMPL-package}
\alias{compriskMPL}
\alias{compriskMPL-package}
\title{compriskMPL: Penalised Maximum Likelihood for Competing Risk Survival Analysis Models}
\description{
Estimate the regression coefficients and the baseline hazard of proportional hazard Cox models for competing risks data with left, right or interval censored survival data using maximum penalised likelihood. A 'non-parametric' smooth estimate of the baseline hazard function is provided.
}
\keyword{internal}
